backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
yt-dlp>=2025.1.1
openai-whisper==20240930
google-genai
python-dotenv==1.0.1
pydantic==2.10.4
